+++ Src/Sample/Sample.CommandService/Controllers/CommandController.cs
@@ -53,7 +53,7 @@ namespace Sample.CommandService.Controllers
             return Task.Factory.StartNew(() =>
                                              new ApiResult
                                              {
-                                                 ErrorCode = ErrorCode.CommandInvalid,
+                                                 ErrorCode = (int)ErrorCode.CommandInvalid,
                                                  Message = string.Join(",", ModelState.Values
                                                                                       .SelectMany(v => v.Errors
                                                                                                         .Select(e => e.ErrorMessage)))
